Fix some data size warning

+++ clippy/storage.c
@@ -38,7 +38,7 @@ void print_storage() {
     fflush(stdout);
 }
 
-int put_message(int region, unsigned long timestamp, int hash, int len,
+int put_message(int region, unsigned long timestamp, int hash, size_t len,
                 char *buf) {
     if (region > MAX_ELEMENTS - 1) {
         log_error("Naughty tried to put message in excess position");
